Revert "Changes the type of Size() to uint64 (#2074)"
This reverts commit 1458ccc8b07e87f68d3d29c2149ce77679b3e6ea.

+++ imports/wasi_snapshot_preview1/clock_test.go
@@ -259,7 +259,7 @@ func Test_clockTimeGet_Errors(t *testing.T) {
 	mod, r, log := requireProxyModule(t, wazero.NewModuleConfig())
 	defer r.Close(testCtx)
 
-	memorySize := uint32(mod.Memory().Size())
+	memorySize := mod.Memory().Size()
 
 	tests := []struct {
 		name                     string
